+++ code/crates/starknet/host/src/actor.rs
@@ -333,7 +333,10 @@ async fn on_get_value(
     let mut sequence = 0;
 
     while let Some(part) = rx_part.recv().await {
-        state.host.part_store.store(height, round, part.clone());
+        state
+            .host
+            .part_store
+            .store(&stream_id, height, round, part.clone());
 
         debug!(%stream_id, %sequence, "Broadcasting proposal part");
 
@@ -347,7 +350,7 @@ async fn on_get_value(
         sequence += 1;
     }
 
-    let msg = StreamMessage::new(stream_id, sequence, StreamContent::Fin);
+    let msg = StreamMessage::new(stream_id.clone(), sequence, StreamContent::Fin);
     network.cast(NetworkMsg::PublishProposalPart(msg))?;
 
     let block_hash = rx_hash.await?;
@@ -356,14 +359,14 @@ async fn on_get_value(
     state
         .host
         .part_store
-        .store_value_id(height, round, block_hash);
+        .store_value_id(&stream_id, height, round, block_hash);
 
-    let parts = state.host.part_store.all_parts(height, round);
+    let parts = state
+        .host
+        .part_store
+        .all_parts_by_stream_id(stream_id, height, round);
 
-    let Some(value) = state.build_value_from_parts(&parts, height, round).await else {
-        error!(%height, %round, "Failed to build block from parts");
-        return Ok(());
-    };
+    let value = state.build_proposal_from_parts(height, round, &parts).await;
 
     debug!(%height, %round, %block_hash, "Storing proposed value from assembled block");
     if let Err(e) = state.block_store.store_undecided_value(value.clone()).await {
@@ -466,7 +469,10 @@ async fn on_restream_proposal(
             PartType::Fin => fin_part.clone(),
         };
 
-        state.host.part_store.store(height, round, new_part.clone());
+        state
+            .host
+            .part_store
+            .store(&stream_id, height, round, new_part.clone());
 
         debug!(%stream_id, %sequence, "Broadcasting proposal part");
 
@@ -550,10 +556,6 @@ async fn on_received_proposal_part(
     from: PeerId,
     reply_to: RpcReplyPort<ProposedValue<MockContext>>,
 ) -> Result<(), ActorProcessingErr> {
-    // TODO: Use state.host.receive_proposal() and move some of the logic below there
-
-    let sequence = part.sequence;
-
     // When inserting part in a map, stream tries to connect all received parts in the right order,
     // starting from beginning and emits parts sequence chunks  when it succeeds
     // If it can't connect part, it buffers it
@@ -563,9 +565,15 @@ async fn on_received_proposal_part(
     // 2 arrives -> 2, 3 and 4 are emitted
 
     // `insert` returns connected sequence of parts if any is emitted
-    let Some(parts) = state.part_streams_map.insert(from, part) else {
+    // If all parts have been received the stream is removed from the map streams
+    let Some(parts) = state.part_streams_map.insert(from, part.clone()) else {
         return Ok(());
     };
+
+    // The `part` sequence number must be for the first `ProposalPart` in `parts`.
+    // So we start with this sequence and we increment for the debug log.
+    let mut sequence = part.sequence;
+    let stream_id = part.stream_id;
 
     if parts.height < state.height {
         trace!(
@@ -592,11 +600,14 @@ async fn on_received_proposal_part(
         );
 
         if let Some(value) = state
-            .build_value_from_part(parts.height, parts.round, part)
+            .build_value_from_part(&stream_id, parts.height, parts.round, part)
             .await
         {
             debug!(
-                height = %value.height, round = %value.round, block_hash = %value.value,
+                height = %value.height,
+                round = %value.round,
+                block_hash = %value.value,
+                validity = ?value.validity,
                 "Storing proposed value assembled from proposal parts"
             );
 
@@ -610,6 +621,8 @@ async fn on_received_proposal_part(
             reply_to.send(value)?;
             break;
         }
+
+        sequence += 1;
     }
 
     Ok(())
@@ -625,7 +638,10 @@ async fn on_decided(
 ) -> Result<(), ActorProcessingErr> {
     let (height, round) = (certificate.height, certificate.round);
 
-    let mut all_parts = state.host.part_store.all_parts(height, round);
+    let mut all_parts = state
+        .host
+        .part_store
+        .all_parts_by_value_id(&certificate.value_id);
 
     let mut all_txes = vec![];
     for part in all_parts.iter_mut() {
@@ -664,7 +680,7 @@ async fn on_decided(
     // Prune the PartStore of all parts for heights lower than `state.height`
     state.host.part_store.prune(state.height);
 
-    // Store the block
+    // Prune the block store, keeping only the last `max_retain_blocks` blocks
     prune_block_store(state).await;
 
     // Notify the mempool to remove corresponding txs