+++ code/crates/starknet/host/src/host/state.rs
@@ -1,3 +1,4 @@
+use sha3::Digest;
 use std::path::Path;
 use std::sync::Arc;
 use std::time::Duration;
@@ -5,14 +6,14 @@ use std::time::Duration;
 use rand::RngCore;
 use tracing::{debug, error, trace};
 
-use malachitebft_core_types::{Context, Round, Validity};
+use malachitebft_core_types::{Round, Validity};
 use malachitebft_engine::consensus::ConsensusRef;
 use malachitebft_engine::host::ProposedValue;
 use malachitebft_engine::util::streaming::StreamId;
 use malachitebft_starknet_p2p_proto as p2p_proto;
 
 use crate::block_store::BlockStore;
-use crate::host::{Host, StarknetHost};
+use crate::host::StarknetHost;
 use crate::streaming::PartStreamsMap;
 use crate::types::*;
 
@@ -64,60 +65,50 @@ impl HostState {
         StreamId::new(bytes.into())
     }
 
-    #[tracing::instrument(skip_all, fields(%height, %round))]
-    pub async fn build_value_from_parts(
-        &self,
-        parts: &[Arc<ProposalPart>],
-        height: Height,
-        round: Round,
-    ) -> Option<ProposedValue<MockContext>> {
-        let (valid_round, value, proposer, validity) = self
-            .build_proposal_content_from_parts(parts, height, round)
-            .await?;
-
-        Some(ProposedValue {
-            proposer,
-            height,
-            round,
-            valid_round,
-            value,
-            validity,
-        })
-    }
-
     #[allow(clippy::type_complexity)]
     #[tracing::instrument(skip_all, fields(%height, %round))]
-    pub async fn build_proposal_content_from_parts(
+    pub async fn build_proposal_from_parts(
         &self,
-        parts: &[Arc<ProposalPart>],
         height: Height,
         round: Round,
-    ) -> Option<(Round, BlockHash, Address, Validity)> {
-        if parts.is_empty() {
-            return None;
-        }
+        parts: &[Arc<ProposalPart>],
+    ) -> ProposedValue<MockContext> {
+        // We must be here with non-empty `parts`, must have init, fin, commitment and maybe transactions
+        assert!(!parts.is_empty(), "Parts must not be empty");
 
-        let Some(init) = parts.iter().find_map(|part| part.as_init()) else {
-            error!("Part not found: Init");
-            return None;
-        };
+        let init = parts
+            .iter()
+            .find_map(|part| part.as_init())
+            .expect("Init part not found");
 
-        let Some(fin) = parts.iter().find_map(|part| part.as_fin()) else {
-            error!("Part not found: Fin");
-            return None;
-        };
+        let fin = parts
+            .iter()
+            .find_map(|part| part.as_fin())
+            .expect("Fin part not found");
 
-        let Some(_block_info) = parts.iter().find_map(|part| part.as_block_info()) else {
-            error!("Part not found: BlockInfo");
-            return None;
-        };
+        let _block_info = parts
+            .iter()
+            .find_map(|part| part.as_block_info())
+            .expect("BlockInfo part not found");
 
-        let Some(commitment) = parts.iter().find_map(|part| part.as_commitment()) else {
-            error!("Part not found: ProposalCommitment");
-            return None;
-        };
+        let commitment = parts
+            .iter()
+            .find_map(|part| part.as_commitment())
+            .expect("ProposalCommitment part not found");
 
-        let validity = self.verify_proposal_validity(init, fin, commitment).await?;
+        // Collect all transactions from the transaction parts
+        // We expect that the transaction parts are ordered by sequence number but we don't have a way to check
+        // this here, so we just collect them in the order.
+        let transactions: Vec<Transaction> = parts
+            .iter()
+            .filter_map(|part| part.as_transactions())
+            .flat_map(|batch| batch.as_slice().iter().cloned())
+            .collect();
+
+        // Determine the validity of the proposal
+        let validity = self
+            .verify_proposal_validity(fin, commitment, transactions)
+            .await;
 
         let valid_round = init.valid_round;
         if valid_round.is_defined() {
@@ -126,47 +117,46 @@ impl HostState {
 
         trace!(parts.len = %parts.len(), "Building proposal content from parts");
 
-        Some((
+        ProposedValue {
+            proposer: init.proposer,
+            height,
+            round,
             valid_round,
-            fin.proposal_commitment_hash,
-            init.proposer,
+            value: fin.proposal_commitment_hash,
             validity,
-        ))
+        }
     }
 
     async fn verify_proposal_validity(
         &self,
-        init: &ProposalInit,
-        _fin: &ProposalFin,
-        _commitment: &ProposalCommitment,
-    ) -> Option<Validity> {
-        let validators = self.host.validators(init.height).await?;
+        fin: &ProposalFin,
+        commitment: &ProposalCommitment,
+        transactions: Vec<Transaction>,
+    ) -> Validity {
+        let mut hasher = sha3::Keccak256::new();
 
-        if !validators.iter().any(|v| v.address == init.proposer) {
-            error!(proposer = %init.proposer, "No validator found for the proposer");
-            return None;
-        };
-
-        let validator_set = ValidatorSet::new(validators);
-        let proposer = self
-            .ctx
-            .select_proposer(&validator_set, init.height, init.round);
-
-        if proposer.address != init.proposer {
-            error!(
-                height = %init.height,
-                round = %init.round,
-                proposer = %init.proposer,
-                expected = %proposer.address,
-                "Proposer is not the selected proposer for this height and round"
-            );
-
-            return None;
+        for tx in transactions.iter() {
+            hasher.update(tx.hash().as_bytes());
         }
 
-        // TODO: Check that the hash of `commitment` matches `fin.proposal_commitment_hash`
+        let transaction_commitment = Hash::new(hasher.finalize().into());
 
-        Some(Validity::Valid)
+        // TODO: Check that computed transaction_commitment and state_diff_commitment match the ones in the `commitment` and
+        // the propposal commitment hash matches `fin.proposal_commitment_hash`
+        // For now we just check that the hash of transactions matches the transaction commitment in `commitment`
+        // and that the proposal commitment hash matches `fin.proposal_commitment_hash`
+        let valid_proposal = transaction_commitment == commitment.transaction_commitment
+            && transaction_commitment == fin.proposal_commitment_hash;
+
+        if valid_proposal {
+            Validity::Valid
+        } else {
+            error!(
+                "ProposalCommitment hash mismatch: {} != {}",
+                transaction_commitment, fin.proposal_commitment_hash
+            );
+            Validity::Invalid
+        }
     }
 
     #[tracing::instrument(skip_all, fields(
@@ -176,18 +166,24 @@ impl HostState {
     ))]
     pub async fn build_value_from_part(
         &mut self,
+        stream_id: &StreamId,
         height: Height,
         round: Round,
         part: ProposalPart,
     ) -> Option<ProposedValue<MockContext>> {
-        self.host.part_store.store(height, round, part.clone());
+        self.host
+            .part_store
+            .store(stream_id, height, round, part.clone());
 
         if let ProposalPart::Transactions(txes) = &part {
             if self.host.params.exec_time_per_tx > Duration::from_secs(0) {
                 debug!("Simulating tx execution and proof verification");
 
-                // Simulate Tx execution and proof verification (assumes success)
-                // TODO: Add config knob for invalid blocks
+                // Simulate Tx execution. In the real implementation the results of the execution would be
+                // accumulated in some intermediate state structure based on which the proposal commitment
+                // will be computed once all parts are received and checked against the received
+                // `ProposalCommitment` part (e.g. `state_diff_commitment`) and the `proposal_commitment_hash`
+                // in the `Fin` part.
                 let num_txes = txes.len() as u32;
                 let exec_time = self.host.params.exec_time_per_tx * num_txes;
                 tokio::time::sleep(exec_time).await;
@@ -196,7 +192,10 @@ impl HostState {
             }
         }
 
-        let parts = self.host.part_store.all_parts(height, round);
+        let parts = self
+            .host
+            .part_store
+            .all_parts_by_stream_id(stream_id.clone(), height, round);
 
         trace!(
             count = self.host.part_store.blocks_count(),
@@ -204,7 +203,7 @@ impl HostState {
         );
 
         // TODO: Do more validations, e.g. there is no higher tx proposal part,
-        //       check that we have received the proof, etc.
+        // check that we have received the proof, etc.
         let Some(_fin) = parts.iter().find_map(|part| part.as_fin()) else {
             debug!("Proposal part has not been received yet: Fin");
             return None;
@@ -228,14 +227,13 @@ impl HostState {
             "All parts have been received already, building value"
         );
 
-        let result = self.build_value_from_parts(&parts, height, round).await;
+        // TODO: Add config knob for invalid blocks
+        let proposed_value = self.build_proposal_from_parts(height, round, &parts).await;
 
-        if let Some(ref proposed_value) = result {
-            self.host
-                .part_store
-                .store_value_id(height, round, proposed_value.value);
-        }
+        self.host
+            .part_store
+            .store_value_id(stream_id, height, round, proposed_value.value);
 
-        result
+        Some(proposed_value)
     }
 }