+++ code/crates/test/app/src/state.rs
@@ -175,8 +175,6 @@ impl State {
             ));
         };
 
-        self.store.remove_undecided_proposal(height, round).await?;
-
         let middleware = self.ctx.middleware();
         debug!(%height, %round, "Middleware: {middleware:?}");
 
@@ -187,9 +185,9 @@ impl State {
                     .store_decided_value(&certificate, proposal.value)
                     .await?;
 
-                // Prune the store, keep the last HISTORY_LENGTH values
+                // Prune the store, keep the last HISTORY_LENGTH decided values, remove all undecided proposals for the decided height
                 let retain_height = Height::new(height.as_u64().saturating_sub(HISTORY_LENGTH));
-                self.store.prune(retain_height).await?;
+                self.store.prune(height, retain_height).await?;
 
                 // Move to next height
                 self.current_height = self.current_height.increment();
@@ -213,25 +211,32 @@ impl State {
         Ok(())
     }
 
-    /// Retrieves a previously built proposal value for the given height
+    /// Retrieves a previously built proposal value for the given height and round.
+    /// Called by the consensus engine to re-use a previously built value.
+    /// There should be at most one proposal for a given height and round when the proposer is not byzantine.
+    /// We assume this implementation is not byzantine and we are the proposer for the given height and round.
+    /// Therefore there must be a single proposal for the rounds where we are the proposer, with the proposer address matching our own.
     pub async fn get_previously_built_value(
         &self,
         height: Height,
         round: Round,
     ) -> eyre::Result<Option<LocallyProposedValue<TestContext>>> {
-        let Some(proposal) = self.store.get_undecided_proposal(height, round).await? else {
-            return Ok(None);
-        };
+        let proposals = self.store.get_undecided_proposals(height, round).await?;
 
-        Ok(Some(LocallyProposedValue::new(
-            proposal.height,
-            proposal.round,
-            proposal.value,
-        )))
+        assert!(
+            proposals.len() <= 1,
+            "There should be at most one proposal for a given height and round"
+        );
+
+        proposals
+            .first()
+            .map(|p| LocallyProposedValue::new(p.height, p.round, p.value.clone()))
+            .map(Some)
+            .map(Ok)
+            .unwrap_or_else(|| Ok(None))
     }
 
-    /// Creates a new proposal value for the given height
-    /// Returns either a previously built proposal or creates a new one
+    /// Creates a new proposal value for the given height and round
     async fn create_proposal(
         &mut self,
         height: Height,
@@ -240,7 +245,7 @@ impl State {
         assert_eq!(height, self.current_height);
         assert_eq!(round, self.current_round);
 
-        // We create a new value.
+        // Create a new value
         let value = self.make_value();
 
         let proposal = ProposedValue {
@@ -252,7 +257,7 @@ impl State {
             validity: Validity::Valid, // Our proposals are de facto valid
         };
 
-        // Insert the new proposal into the undecided proposals.
+        // Insert the new proposal into the undecided proposals
         self.store
             .store_undecided_proposal(proposal.clone())
             .await?;
@@ -285,7 +290,6 @@ impl State {
     }
 
     /// Creates a new proposal value for the given height
-    /// Returns either a previously built proposal or creates a new one
     pub async fn propose_value(
         &mut self,
         height: Height,
@@ -329,7 +333,11 @@ impl State {
             msgs.push(msg);
         }
 
-        msgs.push(StreamMessage::new(stream_id, sequence, StreamContent::Fin));
+        msgs.push(StreamMessage::new(
+            stream_id.clone(),
+            sequence,
+            StreamContent::Fin,
+        ));
 
         msgs.into_iter()
     }