+++ code/crates/test/app/src/store/keys.rs
@@ -1,9 +1,9 @@
 use core::mem::size_of;
 
 use malachitebft_app_channel::app::types::core::Round;
-use malachitebft_test::Height;
+use malachitebft_test::{Height, ValueId};
 
-pub type UndecidedValueKey = (HeightKey, RoundKey);
+pub type UndecidedValueKey = (HeightKey, RoundKey, ValueIdKey);
 
 #[derive(Copy, Clone, Debug)]
 pub struct HeightKey;
@@ -79,5 +79,43 @@ impl redb::Value for RoundKey {
 impl redb::Key for RoundKey {
     fn compare(data1: &[u8], data2: &[u8]) -> std::cmp::Ordering {
         <i64 as redb::Key>::compare(data1, data2)
+    }
+}
+
+#[derive(Copy, Clone, Debug)]
+pub struct ValueIdKey;
+
+impl redb::Value for ValueIdKey {
+    type SelfType<'a> = ValueId;
+    type AsBytes<'a> = [u8; size_of::<u64>()];
+
+    fn fixed_width() -> Option<usize> {
+        Some(size_of::<u64>())
+    }
+
+    fn from_bytes<'a>(data: &'a [u8]) -> Self::SelfType<'a>
+    where
+        Self: 'a,
+    {
+        let id = <u64 as redb::Value>::from_bytes(data);
+        ValueId::new(id)
+    }
+
+    fn as_bytes<'a, 'b: 'a>(value: &'a Self::SelfType<'b>) -> Self::AsBytes<'a>
+    where
+        Self: 'a,
+        Self: 'b,
+    {
+        <u64 as redb::Value>::as_bytes(&value.as_u64())
+    }
+
+    fn type_name() -> redb::TypeName {
+        redb::TypeName::new("ValueId")
+    }
+}
+
+impl redb::Key for ValueIdKey {
+    fn compare(data1: &[u8], data2: &[u8]) -> std::cmp::Ordering {
+        <u64 as redb::Key>::compare(data1, data2)
     }
 }